+++ server/container_attach.go
@@ -3,25 +3,13 @@ package server
 import (
 	"fmt"
 	"io"
-	"net"
-	"os"
-	"path/filepath"
 	"time"
 
 	"github.com/kubernetes-sigs/cri-o/oci"
-	"github.com/kubernetes-sigs/cri-o/utils"
 	"github.com/sirupsen/logrus"
 	"golang.org/x/net/context"
 	"k8s.io/client-go/tools/remotecommand"
 	pb "k8s.io/kubernetes/pkg/kubelet/apis/cri/runtime/v1alpha2"
-	kubecontainer "k8s.io/kubernetes/pkg/kubelet/container"
-)
-
-/* Sync with stdpipe_t in conmon.c */
-const (
-	AttachPipeStdin  = 1
-	AttachPipeStdout = 2
-	AttachPipeStderr = 3
 )
 
 // Attach prepares a streaming endpoint to attach to a running container.
@@ -57,96 +45,5 @@ func (ss streamService) Attach(containerID string, inputStream io.Reader, output
 		return fmt.Errorf("container is not created or running")
 	}
 
-	controlPath := filepath.Join(c.BundlePath(), "ctl")
-	controlFile, err := os.OpenFile(controlPath, os.O_WRONLY, 0)
-	if err != nil {
-		return fmt.Errorf("failed to open container ctl file: %v", err)
-	}
-	defer controlFile.Close()
-
-	kubecontainer.HandleResizing(resize, func(size remotecommand.TerminalSize) {
-		logrus.Debugf("Got a resize event: %+v", size)
-		_, err := fmt.Fprintf(controlFile, "%d %d %d\n", 1, size.Height, size.Width)
-		if err != nil {
-			logrus.Debugf("Failed to write to control file to resize terminal: %v", err)
-		}
-	})
-
-	attachSocketPath := filepath.Join(ss.runtimeServer.Config().ContainerAttachSocketDir, c.ID(), "attach")
-	conn, err := net.DialUnix("unixpacket", nil, &net.UnixAddr{Name: attachSocketPath, Net: "unixpacket"})
-	if err != nil {
-		return fmt.Errorf("failed to connect to container %s attach socket: %v", c.ID(), err)
-	}
-	defer conn.Close()
-
-	receiveStdout := make(chan error)
-	if outputStream != nil || errorStream != nil {
-		go func() {
-			receiveStdout <- redirectResponseToOutputStreams(outputStream, errorStream, conn)
-		}()
-	}
-
-	stdinDone := make(chan error)
-	go func() {
-		var err error
-		if inputStream != nil {
-			_, err = utils.CopyDetachable(conn, inputStream, nil)
-			conn.CloseWrite()
-		}
-		stdinDone <- err
-	}()
-
-	select {
-	case err := <-receiveStdout:
-		return err
-	case err := <-stdinDone:
-		if _, ok := err.(utils.DetachError); ok {
-			return nil
-		}
-		if outputStream != nil || errorStream != nil {
-			return <-receiveStdout
-		}
-	}
-
-	return nil
-}
-
-func redirectResponseToOutputStreams(outputStream, errorStream io.Writer, conn io.Reader) error {
-	var err error
-	buf := make([]byte, oci.BufSize+1)
-
-	for {
-		nr, er := conn.Read(buf)
-		if nr > 0 {
-			var dst io.Writer
-			if buf[0] == AttachPipeStdout {
-				dst = outputStream
-			} else if buf[0] == AttachPipeStderr {
-				dst = errorStream
-			} else {
-				logrus.Debugf("Got unexpected attach type %+d", buf[0])
-			}
-
-			if dst != nil {
-				nw, ew := dst.Write(buf[1:nr])
-				if ew != nil {
-					err = ew
-					break
-				}
-				if nr != nw+1 {
-					err = io.ErrShortWrite
-					break
-				}
-			}
-		}
-		if er == io.EOF {
-			break
-		}
-		if er != nil {
-			err = er
-			break
-		}
-	}
-
-	return err
+	return ss.runtimeServer.Runtime().AttachContainer(c, inputStream, outputStream, errorStream, tty, resize)
 }
